+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All Notable changes to `laravel-console` will be documented in this file
 
+## NEXT - 2015-12-28
+
+### Added
+- add Admin LTE backbone
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- Nothing
+
+### Security
+- Nothing
+
 ## NEXT - 2015-12-12
 
 ### Added